% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{aux_equal_Int}
\alias{aux_equal_Int}
\title{This function returns the pointwise interval with credibility level of 100(1 - alpha)\% and equal tails (each tail has alpha/2 area).}
\usage{
aux_equal_Int(post_matrix, alpha)
}
\description{
This function returns the pointwise interval with credibility level of 100(1 - alpha)\% and equal tails (each tail has alpha/2 area).
}
