% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{aux_hpd_Int}
\alias{aux_hpd_Int}
\title{This function returns the pointwise HPD interval with credibility level of 100(1 - alpha)\%.}
\usage{
aux_hpd_Int(post_matrix, alpha)
}
\description{
This function returns the pointwise HPD interval with credibility level of 100(1 - alpha)\%.
}
