% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B1.R
\name{band_B1}
\alias{band_B1}
\title{B1 band}
\usage{
band_B1(post, alpha, interval)
}
\arguments{
\item{post}{Matrix with dimension n.time by n.size containing posterior values for n.time times and n.size posterior sample.}

\item{alpha}{Value inside the interval (0,1) in order to get a credibility level of 100(1 - alpha)\%.}

\item{interval}{Choice between "equal" and "hpd". The "equal" considers equal tails (each tail has alpha/2 area) and "hpd" considers the HPD interval.}
}
\value{
A list with the following components:
\item{B1}{A matrix with n.time rows and 2 columns with B1 band.}
\item{alpha}{The information about the credibility level of 100(1 - alpha)\%.}
\item{interval}{If "equal" or "hpd".}
}
\description{
\code{band_B1}  return a matrix with B1 band.
}
