% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2.R
\name{band_B2}
\alias{band_B2}
\title{B2 band}
\usage{
band_B2(post, alpha, density_aux)
}
\arguments{
\item{post}{Matrix with dimension n.time by n.size containing posterior values for n.time times and n.size posterior sample.}

\item{alpha}{Value inside the interval (0,1) in order to get a credibility level of 100(1 - alpha)\%.}

\item{density_aux}{Posterior density evaluated in each MCMC values of the parameters.}
}
\value{
A list with the following components:
\item{B2}{A matrix with n.time rows and 2 columns with B2 band.}
\item{alpha}{The information about the credibility level of 100(1 - alpha)\%.}
}
\description{
\code{band_B2}  return a matrix with B2 band.
}
