% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointwise.R
\name{band_PW}
\alias{band_PW}
\title{Pointwise band}
\usage{
band_PW(post, alpha, interval)
}
\arguments{
\item{post}{Matrix with dimension n.time by n.size containing posterior values for n.time times and n.size posterior sample.}

\item{alpha}{Value inside the interval (0,1) in order to get a credibility level of 100(1 - alpha)\%.}

\item{interval}{Choice between "equal" and "hpd". The "equal" considers equal tails (each tail has alpha/2 area) and "hpd" considers the HPD interval.}
}
\value{
A list with the following components:
\item{PW}{A matrix with n.time rows and 2 columns with pointwise band.}
\item{alpha}{The information about the credibility level of 100(1 - alpha)\%.}
}
\description{
\code{band_PW}  return a matrix with pointwise band.
}
